Add feature: install only dependencies of a package (#9)
This PR allows installing some specs fully, and some specs only the dependencies. This is useful when you want to create a base image for your software for which you have a spack package.py. Then dependencies will be picked up automatically by spack, and the information does not have to be duplicated in the Dockerfile and in the package.py.